my_endpoint/pyproject.toml
[project]
name = "my-endpoint"
version = "0.1.0"
description = ""
authors = [
    {name = "Crambit",email = "yashjuly20@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "ag-ui-protocol (>=0.1.5,<0.2.0)",
    "openai (>=1.82.1,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)"
]

[tool.poetry]
packages = [{include = "my_endpoint", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
